Change float representations to be more IEEE-ish
Implemented BitsToReal WRONGLY for like 6 hours.
It's probably still wrong. It works for 8 bit (2,5) floats but probably only because I got lucky.

I am a bit shit at bit shifting.

+++ src/real.h
@@ -39,7 +39,7 @@ namespace IPDF
 	// Mostly in the testers.
 	inline float Float(float f) {return f;}
 	inline double Float(double f) {return f;}
-	inline long double Float(long double f) {return f;}
+	inline double Float(long double f) {return double(f);}
 }
 
 #endif //_REAL_H
